% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crostonsdecomp}
\alias{crostonsdecomp}
\title{Croston's decomposition}
\usage{
crostonsdecomp(y)
}
\arguments{
\item{y}{A vector of data.}
}
\value{
The function return a list of 3 series:
\describe{
\item{occurrence}{The binarised series between positive and negative values.}
\item{demand}{The subseries of positive values.}
\item{intervals}{A sequence of time intervals between positive observations.}
}
}
\description{
Divide a time series into occurrence and demand.
}
