% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{epsilon}
\alias{epsilon}
\title{Internal epsilon value for nloptr constraints}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
epsilon
}
\description{
Internal epsilon value for nloptr constraints
}
\keyword{internal}
