% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_nbinom}
\alias{fit_nbinom}
\title{Negative binomial MLEs}
\usage{
fit_nbinom(y)
}
\arguments{
\item{y}{A vector of data.}
}
\value{
A vector containing \code{size} and \code{prob} parameters
}
\description{
Fit the distribution using an optimiser.
}
