% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_forecast}
\alias{get_forecast}
\title{Produce the forecasts from the samples.}
\usage{
get_forecast(forecast_samples, levels, cumulative, side, h)
}
\arguments{
\item{forecast_samples}{An array containing predictive sample paths.}

\item{levels}{The confidence levels for predictive intervals.}

\item{cumulative}{Whether to just return the sum over the forecast horizon.}

\item{side}{The side of the confidence intervals.}

\item{h}{The number of steps ahead used as forecasting horizon.}
}
\value{
A list with the following components:
\describe{
\item{mean_forecast}{The predicted mean.}
\item{prob_zero}{The predicted probability of observing zero.}
\item{levels_upper}{The upper confidence levels of the interval.}
\item{levels_lower}{The lower confidence levels of the interval.}
\item{CI_upper}{The upper bounds of the confidence region.}
\item{CI_lower}{The lower bounds of the confidence region.}
}
}
\description{
This functions generates mean forecast with confidence intervals
and predicts the probabilities of zero starting from samples.
It is intended for internal use only and is not exported.
}
\details{
This function is used within other package functions.
}
\keyword{internal}
