% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_data}
\alias{init_data}
\title{Initialize the data.}
\usage{
init_data(data, holdout, h)
}
\arguments{
\item{data}{A time series or a numeric vector or 1-dimensional array.}

\item{holdout}{Whether to remove the last h values from the in-sample part.}

\item{h}{The number of steps ahead used as forecasting horizon.}
}
\value{
A list with the following components:
\describe{
\item{y_insample}{The in-sample part of the data.}
\item{y_holdout}{The holdout part of the data.}
\item{freq}{The time frequency.}
\item{pred_start}{The starting time for the predictions.}
\item{L}{The length of the in-sample part.}
}
}
\description{
This function extrapolates in-sample, holdout, frequency, starting time and
length from a time series. It is intended for internal use only and is not exported.
}
\details{
This function is used within other package functions.
}
\keyword{internal}
