% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticd.R
\name{staticd}
\alias{staticd}
\title{Static parametric distribution.}
\usage{
staticd(
  data,
  h = 10,
  levels = 0.9,
  holdout = FALSE,
  cumulative = FALSE,
  side = c("upper", "both", "lower"),
  nsim = 10000,
  paths = FALSE,
  rmzeros = FALSE,
  distr = c("auto", "pois", "zip", "nbinom", "zinb", "mixture")
)
}
\arguments{
\item{data}{A time series or a numeric vector or 1-dimensional array.}

\item{h}{The number of steps ahead used as forecasting horizon.}

\item{levels}{The confidence levels for predictive intervals.}

\item{holdout}{Whether to remove the last h values from the in-sample part.}

\item{cumulative}{Whether to just return the sum over the forecast horizon.}

\item{side}{The side of the confidence intervals.}

\item{nsim}{The amount of predictive sample paths.}

\item{paths}{Whether to return sample trajectories.}

\item{rmzeros}{Whether to remove the initial zeros from the series.}

\item{distr}{The distribution choice to be fit.}
}
\value{
The function returns a model of a class "#TODO", which contains:
\describe{
\item{model}{The model name.}
\item{call}{The arguments of the function call.}
\item{params}{Internal parameters for the estimation of the model.}
\item{forecast}{The predicted mean.}
\item{probzero}{The predicted probability of observing zero.}
\item{levels_upper}{The upper confidence levels of the interval.}
\item{levels_lower}{The lower confidence levels of the interval.}
\item{upperCI}{The upper bounds of the confidence region.}
\item{lowerCI}{The lower bounds of the confidence region.}
\item{samples}{The predictive sample paths.}
}
}
\description{
Maximise the marginal likelihood of the time series using distributions as Poisson,
negative binomial or their zero-inflated versions.
}
\details{
Short description #TODO.
}
\examples{
y <- c(rpois(30,0.9), rpois(20,0.3))
staticd <- staticd(y)

}
\references{
\itemize{
\item paper to cite # TODO
}
}
\seealso{
\code{\link[package]{function}, \link[package2]{function2}} #TODO

Other models: 
\code{\link{gampois}()},
\code{\link{seam}()},
\code{\link{wss}()}
}
\author{
Stefano Damato, \email{stefano.damato@idsia.ch}
}
\concept{models}
\keyword{#TODO}
\keyword{some}
\keyword{words}
